requirements.txt
scikit-image==0.25.2
pytubefix==8.12.2
opencv-python==4.11.0.86
pillow==11.2.1
customtkinter==5.2.2